bring diauxic plot to txp_mtb_ge compartment, remove glc_lcts_shifter
